+++ src/components/dashboards/admin/adminDashboard.js
@@ -40,10 +40,6 @@ const AdminDashboard = () => {
   const [punishmentData, setPunishmentData] = useState([]);
   const [modalType, setModalType] = useState("");
   const [adminDto, setAdminDto] = useState([]);
-  const [activeIndex, setActiveIndex] = useState(null);
-  const [updatePage, setUpdatePage] = useState(false);
-  const [referralList, setReferralList] = useState([null]);
-  const [deletePayload, setDeletePayload] = useState(null);
   const [openModal, setOpenModal] = useState({
     display: false,
     message: "",
@@ -51,6 +47,7 @@ const AdminDashboard = () => {
     data: null,
   });
   const [textareaValue, setTextareaValue] = useState("");
+  const [referralList, setReferralList] = useState([null]);
   const [toast, setToast] = useState({ visible: false, message: "" });
   const [loadingPunishmentId, setLoadingPunishmentId] = useState({
     id: null,
@@ -79,33 +76,8 @@ const AdminDashboard = () => {
     }
   }, []);
 
-  const handleUpdatePage = () => {
-    setTimeout(() => {
-      setUpdatePage((prev) => !prev);
-    }, 500);
-  };
-
-  const handleClosePunishment = (description, id) => {
-    const payload = { id: id, comment: description };
-
-    const url = `${baseUrl}/officeReferral/v1/closeId`;
-    axios
-      .post(url, payload, { headers })
-      .then((response) => {
-        console.log(response.data);
-        handleUpdatePage();
-      })
-      .catch((error) => {
-        console.error(error);
-      });
-  };
-
   const toggleNotificationDrawer = (open) => {
     setOpenNotificationDrawer(open);
-  };
-
-  const handleTextareaChange = (event) => {
-    setTextareaValue(event.target.value);
   };
 
   //Fetch Data to Prop Drill to Componetns
@@ -127,33 +99,6 @@ const AdminDashboard = () => {
     }
   }, [panelName]);
 
-  const handleRejectPunishment = (obj) => {
-    setLoadingPunishmentId({
-      id: obj.officeReferral.officeReferralId,
-      buttonType: "close",
-    });
-    const url = `${baseUrl}/officeReferral/v1/rejected/${obj.officeReferral.officeReferralId}`;
-    axios
-      .put(url, [textareaValue], { headers }) // Pass the headers option with the JWT token
-      .then(function (response) {
-        setToast({
-          visible: true,
-          message:
-            "You have rejected the student's answers and an email has been sent letting them know.",
-        });
-      })
-      .catch(function (error) {
-        console.error(error);
-      })
-      .finally(() => {
-        setOpenModal({ display: false, message: "" });
-        setTimeout(() => {
-          setToast({ visible: false, message: "" });
-          setLoadingPunishmentId({ id: null, buttonType: "" });
-        }, 1000);
-      });
-  };
-
   return (
     loggedIn && (
       <div>
@@ -163,103 +108,6 @@ const AdminDashboard = () => {
               setContactUsDisplayModal={setModalType}
               contactUsDisplayModal={undefined}
             />
-          )}
-          {openModal.display && (
-            <div className="modal-overlay">
-              <div className="modal-content">
-                <div className="modal-header">
-                  <h3>{openModal.message}</h3>
-                  <div className="answer-container">
-                    {openModal.data.officeReferral.infractionDescription.map(
-                      (item, index) => {
-                        if (index > 1) {
-                          const match = item.match(
-                            /question=([\s\S]+?),\s*answer=([\s\S]+?)(?=\))/
-                          );
-                          if (match) {
-                            const question = match[1].trim();
-                            const answer = match[2].trim();
-
-                            return (
-                              <div
-                                key={index}
-                                style={{
-                                  display: "flex",
-                                  flexDirection: "row",
-                                  border: "1px solid black",
-                                }}
-                              >
-                                <div
-                                  style={{
-                                    backgroundColor: "grey",
-                                    minHeight: "15px",
-                                    width: "40%",
-                                  }}
-                                >
-                                  <strong>Question:</strong> {question}
-                                </div>
-                                <div
-                                  style={{
-                                    color: "black",
-                                    backgroundColor: "lightBlue",
-                                    minHeight: "50px",
-                                    width: "60%",
-                                    textAlign: "left",
-                                    paddingLeft: "10px",
-                                  }}
-                                >
-                                  <strong>Answer:</strong> {answer}
-                                </div>
-                              </div>
-                            );
-                          }
-                        }
-                      }
-                    )}
-                  </div>
-                </div>
-                <div className="modal-body">
-                  <textarea
-                    value={textareaValue} // Set the value of the textarea to the state variable
-                    onChange={handleTextareaChange} // Handle changes to the textarea
-                    className="multi-line-input"
-                    placeholder="Enter additional comments"
-                    rows={4} // This sets the initial height to show 4 rows
-                  ></textarea>
-                </div>
-                <div className="modal-buttons">
-                  <button
-                    onClick={() => {
-                      setOpenModal({ display: false, message: "" });
-                      setTextareaValue("");
-                    }}
-                  >
-                    Cancel
-                  </button>
-                  <button
-                    disabled={textareaValue.length === 0}
-                    style={{
-                      backgroundColor: textareaValue === "" ? "grey" : "red",
-                    }}
-                    onClick={() => handleRejectPunishment(deletePayload)}
-                  >
-                    Reject Answers
-                  </button>
-                  <button
-                    disabled={textareaValue.length === 0}
-                    style={{
-                      backgroundColor: textareaValue === "" ? "grey" : "green",
-                    }}
-                    onClick={() => {
-                      handleClosePunishment(deletePayload);
-                      console.log(deletePayload);
-                    }}
-                  >
-                    Accept Answers
-                  </button>
-                </div>
-              </div>
-            </div>
           )}
 
           <Snackbar
@@ -295,126 +143,7 @@ const AdminDashboard = () => {
                   <LoadingWheelPanel />
                 ) : (
                   panelName === "overview" && (
-                    <>
-                      <div className="office-referral-first">
-                        {referralList.map((item, index) => {
-                          return (
-                            <div
-                              className="task-card"
-                              onClick={() => setActiveIndex(index)}
-                              key={index}
-                              role="button"
-                            >
-                              <div className="tag">
-                                <div className="color-stripe"></div>
-                                <div className="tag-content">
-                                  <div className="index"> {index + 1}</div>
-                                  <div className="date">
-                                    {" "}
-                                    {dateCreateFormat(item?.timeCreated)}
-                                  </div>
-                                </div>
-                              </div>
-
-                              <div className="card-body">
-                                <div className="card-body-title">
-                                  {item?.referralCode.codeName}
-                                </div>
-                                <div className="card-body-description">
-                                  {item?.referralDescription[0]}
-                                </div>
-                                {/* USE THIS TO MAKE BADGES FOR OFFICE REFERRALS
-                               {item.referralDescription &&
-                          item.referralDescription[0] !== undefined &&
-                          categoryBadgeGenerator(item.referralDescription[0])} */}
-                              </div>
-                              <div className="card-body">
-                                <div className="card-body-title">
-                                  Created By
-                                </div>
-                                <div className="card-body-description">
-                                  {item?.teacherEmail}
-                                </div>
-                              </div>
-                              <div className="card-actions">
-                                <button
-                                  style={{
-                                    height: "60px",
-                                    width: "180px",
-                                    backgroundColor: "green",
-                                  }}
-                                  onClick={() => {
-                                    setOpenModal({
-                                      display: true,
-                                      message:
-                                        "You are attempting to remove the restorative assignment and close out a referral. If this was not your intent click cancel. If this is your intent, provide a brief explanation for why the restorative assignment is being removed and click Close",
-                                      buttonType: "close",
-                                    });
-                                    setDeletePayload(item);
-                                  }}
-                                >
-                                  <p
-                                    style={{
-                                      marginBottom: "5px",
-                                      marginTop: "-2%",
-                                    }}
-                                  >
-                                    Close Referral
-                                  </p>
-                                  {loadingPunishmentId.id ===
-                                    item.officeReferralId &&
-                                  loadingPunishmentId.buttonType === "close" ? (
-                                    <CircularProgress
-                                      style={{ height: "20px", width: "20px" }}
-                                      color="secondary"
-                                    />
-                                  ) : (
-                                    <CheckBoxIcon />
-                                  )}
-                                </button>
-                                <button
-                                  style={{
-                                    height: "60px",
-                                    width: "180px",
-                                    backgroundColor: "red",
-                                  }}
-                                  onClick={() => {
-                                    setOpenModal({
-                                      display: true,
-                                      message:
-                                        "You are attempting to delete the record of this referral. If you were attempting to remove the restorative assignment and close out the referral please click cancel and hit the “Close Referral” button. If you still want to delete the record of this referral, provide a brief explanation for this action and click Delete Referral.",
-                                      buttonType: "delete",
-                                    });
-                                    setDeletePayload(item);
-                                  }}
-                                >
-                                  <p
-                                    style={{
-                                      marginBottom: "5px",
-                                      marginTop: "-2%",
-                                    }}
-                                  >
-                                    Delete Referral
-                                  </p>
-                                  {loadingPunishmentId.id ===
-                                    item.officeReferralId &&
-                                  loadingPunishmentId.buttonType ===
-                                    "delete" ? (
-                                    <CircularProgress
-                                      style={{ height: "20px", width: "20px" }}
-                                      color="secondary"
-                                    />
-                                  ) : (
-                                    <DeleteForeverIcon />
-                                  )}
-                                </button>
-                              </div>
-                            </div>
-                          );
-                        })}
-                      </div>
                       <AdminOverviewPanel data={adminDto} />
-                    </>
                   )
                 )}
                 {panelName === "viewTeacher" && <AdminTeacherPanel />}
